+++ dotnet/src/SmartAmbience/Threading/AmbienceHub.cs
@@ -1,17 +1,25 @@
 ﻿using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Text;
+using System.Text.RegularExpressions;
+using System.Xml.Linq;
+using System.Threading;
+using System;
 
-namespace System.Threading {
+namespace DistributedDataFlow {
 
-  public static class AmbienceHub {
+  public static partial class AmbienceHub {
+
+    internal delegate void CaptureMethodWithContractInfo(FlowingContractDefinition contract, Action<string, string> target);
+    internal delegate void RestoreMethodWithContractInfo(FlowingContractDefinition contract, IEnumerable<KeyValuePair<string, string>> source);
 
     [DebuggerDisplay("{EndpointName}")]
     private class EndpointAdapter {
       public String EndpointName { get; set; }
-      public Action<Action<string, string>> CaptureMethod { get; set; }
-      public Action<IEnumerable<KeyValuePair<string, string>>> RestoreMethod { get; set; }
+      public CaptureMethodWithContractInfo CaptureMethod { get; set; }
+      public RestoreMethodWithContractInfo RestoreMethod { get; set; }
       public int AssertRestorePerformanceMs { get; set; }
       public string[] RestoreAfter { get; set; }
     }
@@ -29,7 +37,7 @@ namespace System.Threading {
     /// customized/empty set of items to the endpoints OR
     /// just set the OnRestoreNullHandler=null to skip any restore silently.
     /// Note: this quite different from calling each endpoint with an empty set of entries
-    /// - here the endpoints WONT GET ANY TRIGGER!!!
+    /// - in the first case the endpoints WONT GET ANY TRIGGER!!!
     /// </summary>
     public static OnRestoreNullHandlerMethod OnRestoreNullHandler = (
       (ref IDictionary<string, string> defaultsToUse) => {
@@ -66,7 +74,37 @@ namespace System.Threading {
       string endpointName,
       Action<Action<string, string>> captureMethod,
       Action<IEnumerable<KeyValuePair<string, string>>> restoreMethod,
-      int assertRestorePerformanceMs = 500,
+      int assertRestorePerformanceMs = 200,
+      params string[] restoreAfter
+    ) {
+
+      BindCustomEndpoint(
+        endpointName,     
+        (contractName, target) => captureMethod.Invoke(target),
+        (contractName, source) => restoreMethod.Invoke(source),
+        assertRestorePerformanceMs,
+        restoreAfter
+      );
+
+    }
+
+    /// <summary>
+    /// Created a wire-up to a custom endpoint which is participating each time an ambient-snapshot is created or restored.
+    /// </summary>
+    /// <param name="endpointName">
+    /// Note: prepending an "^" befor the name will treat the endpoint as ROOT.
+    /// In this case the name will NOT be added as prefix to the fieldnames when transporting them.
+    /// Any prefixing needs to be dpne by the endpoint itself.
+    /// </param>
+    /// <param name="captureMethod"></param>
+    /// <param name="restoreMethod"></param>
+    /// <param name="assertRestorePerformanceMs"></param>
+    /// <param name="restoreAfter">one or more endpointName on which the endpoint that is currently registered will rely on</param>
+    internal static void BindCustomEndpoint(
+      string endpointName,
+      CaptureMethodWithContractInfo captureMethod,
+      RestoreMethodWithContractInfo restoreMethod,
+      int assertRestorePerformanceMs = 200,
       params string[] restoreAfter
     ) {
 
@@ -101,16 +139,65 @@ namespace System.Threading {
       (string endpointName, int elapsedMs) => Trace.TraceWarning($"{nameof(AmbienceHub)}: Restore performance assert failed for endpoint '{endpointName}'! Restore took {elapsedMs}ms!")
     );
 
+    /// <summary>
+    /// WARNING: THIS IS ONLY MADE TO BE USED FROM ABSTRACT TRANSPORT-TECHNOLOGY
+    /// </summary>
+    /// <param name="targetContainer">
+    /// Container to which a Snapshot of ambient data should be added. 
+    /// Existing entries will be overwritten on key collition, other remaining entries will be preserved.
+    /// </param>
     public static void CaptureCurrentValuesTo(IDictionary<string, string> targetContainer) {
-      CaptureCurrentValuesTo((string key, string value) => targetContainer[key] = value);
+      //NOTE: this overload-method is needed because if we would use an optional parameter flowingContractName=null
+      //in the called method, if wont be compatible for the usage as delegate in the way will be done in some usecases
+      CaptureCurrentValuesTo(targetContainer, null);
     }
 
-    public static void CaptureCurrentValuesTo(Action<string,string> capturingCallback) {
+    /// <summary>
+    /// WARNING: THIS IS ONLY MADE TO BE USED FROM ABSTRACT TRANSPORT-TECHNOLOGY
+    /// </summary>
+    /// <param name="targetContainer">
+    /// Container to which a Snapshot of ambient data should be added. 
+    /// Existing entries will be overwritten on key collition, other remaining entries will be preserved.
+    /// </param>
+    /// <param name="flowingContractName">
+    /// Name of a 'FlowingContract' (optional Feature - see the 'DefineFlowingContract(..)' method)
+    /// which shall be used to select an explicitely defined subset of custom Endpoints and/or 
+    /// AmbientFields to be included in the snapshot.
+    /// </param>
+    public static void CaptureCurrentValuesTo(IDictionary<string, string> targetContainer, string flowingContractName) {
+      CaptureCurrentValuesTo((string key, string value) => targetContainer[key] = value, flowingContractName);
+    }
+
+    /// <summary>
+    /// WARNING: THIS IS ONLY MADE TO BE USED FROM ABSTRACT TRANSPORT-TECHNOLOGY
+    /// </summary>
+    /// <param name="capturingCallback">Callback for processing an Snapshot of ambient data</param>
+    public static void CaptureCurrentValuesTo(Action<string, string> capturingCallback) {
+      //NOTE: this overload-method is needed because if we would use an optional parameter flowingContractName=null
+      //in the called method, if wont be compatible for the usage as delegate in the way will be done in some usecases
+      CaptureCurrentValuesTo(capturingCallback, null);
+    }
+
+    /// <summary>
+    /// WARNING: THIS IS ONLY MADE TO BE USED FROM ABSTRACT TRANSPORT-TECHNOLOGY
+    /// </summary>
+    /// <param name="capturingCallback">Callback for processing an Snapshot of ambient data</param>
+    /// <param name="flowingContractName">
+    /// Name of a 'FlowingContract' (optional Feature - see the 'DefineFlowingContract(..)' method)
+    /// which shall be used to select an explicitely defined subset of custom Endpoints and/or 
+    /// AmbientFields to be included in the snapshot.
+    /// </param>
+    public static void CaptureCurrentValuesTo(Action<string,string> capturingCallback, string flowingContractName) {
       var alreadyCapturedKeys = new HashSet<string>();
 
-      foreach (EndpointAdapter endpoint in _EndpointAdaptersInRestoreOrder) {
+      FlowingContractDefinition contract = AmbienceHub.GetFlowingContractDefinition(flowingContractName);
+
+      IEnumerable<EndpointAdapter> includedEndpoints = _EndpointAdaptersInRestoreOrder.Where((ep) => contract.IsEndpointIncluded(ep.EndpointName));
+
+      foreach (EndpointAdapter endpoint in includedEndpoints) {
         bool isGlobalEndpoint = endpoint.EndpointName.StartsWith("^");
         endpoint.CaptureMethod.Invoke(
+          contract,
           (string key, string value) => {
 
             if (!alreadyCapturedKeys.Add(key)) {
@@ -133,11 +220,14 @@ namespace System.Threading {
                 );
               }
 
+              ////// CAPTURE //////
               capturingCallback.Invoke(key, value);
             }
             else {
-              //add the endpoint name as prefix
+
+              ////// CAPTURE //////
               capturingCallback.Invoke(endpoint.EndpointName + "." + key, value);
+              //                              ^^^ add the endpoint name as prefix
             }
           }
         );
@@ -145,17 +235,41 @@ namespace System.Threading {
 
     }
 
-    public static string CaptureCurrentValuesAsDump(){
+    /// <summary>
+    /// Captures ambient values and dumps them into human readable string, that can be used for Troubleshooting.
+    /// </summary>
+    /// <param name="flowingContractName">
+    /// Name of a 'FlowingContract' (optional Feature - see the 'DefineFlowingContract(..)' method)
+    /// which shall be used to select an explicitely defined subset of custom Endpoints and/or 
+    /// AmbientFields to be included in the snapshot.
+    /// </param>
+    /// <returns>an human readable Snapshot of ambient data</returns>
+    public static string CaptureCurrentValuesAsDump(string flowingContractName = null) {
       var sb = new StringBuilder(2000);
-      CaptureCurrentValuesTo((string key, string value) => sb.AppendLine(key + ": " + value));
+      CaptureCurrentValuesTo((string key, string value) => sb.AppendLine(key + ": " + value), flowingContractName);
       return sb.ToString();
     }
 
-   /// <summary>
-   /// WARNING: THIS IS ONLY MADE TO BE USED FROM ABSTRACT TRANSPORT-TECHNOLOGY
-   /// </summary>
-   /// <param name="sourceToRestore"></param>
-   public static void RestoreValuesFrom(IEnumerable<KeyValuePair<string, string>> sourceToRestore) {
+    /// <summary>
+    /// WARNING: THIS IS ONLY MADE TO BE USED FROM ABSTRACT TRANSPORT-TECHNOLOGY
+    /// </summary>
+    /// <param name="sourceToRestore">an Snapshot of ambient data</param>
+    public static void RestoreValuesFrom(IEnumerable<KeyValuePair<string, string>> sourceToRestore) {
+      //NOTE: this overload-method is needed because if we would use an optional parameter flowingContractName=null
+      //in the called method, if wont be compatible for the usage as delegate in the way will be done in some usecases
+      RestoreValuesFrom(sourceToRestore, null);
+    }
+
+    /// <summary>
+    /// WARNING: THIS IS ONLY MADE TO BE USED FROM ABSTRACT TRANSPORT-TECHNOLOGY
+    /// </summary>
+    /// <param name="sourceToRestore">an Snapshot of ambient data</param>
+    /// <param name="flowingContractName">
+    /// Name of a 'FlowingContract' (optional Feature - see the 'DefineFlowingContract(..)' method)
+    /// which shall be used to select an explicitely defined subset of custom Endpoints and/or 
+    /// AmbientFields to be restored from the given snapshot.
+    /// </param>
+    public static void RestoreValuesFrom(IEnumerable<KeyValuePair<string, string>> sourceToRestore, string flowingContractName) {
 
       if(sourceToRestore == null) {
         IDictionary<string, string> defaultsToRestore = new Dictionary<string, string>();
@@ -170,7 +284,11 @@ namespace System.Threading {
         sourceToRestore = defaultsToRestore;
       }
 
-      foreach (EndpointAdapter endpoint in _EndpointAdaptersInRestoreOrder) {
+      FlowingContractDefinition contract = AmbienceHub.GetFlowingContractDefinition(flowingContractName);
+
+      IEnumerable<EndpointAdapter> includedEndpoints = _EndpointAdaptersInRestoreOrder.Where((ep) => contract.IsEndpointIncluded(ep.EndpointName));
+
+      foreach (EndpointAdapter endpoint in includedEndpoints) {
 
         bool isGlobalEndpoint = endpoint.EndpointName.StartsWith("^");
         var filteredSource = sourceToRestore;
@@ -199,7 +317,8 @@ namespace System.Threading {
 
         DateTime restoreStart = DateTime.Now;
 
-        endpoint.RestoreMethod.Invoke(sourceEntriesForCurrentEndpoint);
+        ////// RESTORE //////
+        endpoint.RestoreMethod.Invoke(contract, sourceEntriesForCurrentEndpoint);
 
         int totalMs = Convert.ToInt32(DateTime.Now.Subtract(restoreStart).TotalMilliseconds);
         if (OnRestorePerformanceAssertFailedMethod != null && totalMs > endpoint.AssertRestorePerformanceMs) {
@@ -210,41 +329,53 @@ namespace System.Threading {
 
     }
 
-    #region " default endpoint for exposed AmbientFields "
+    #region " 'FlowingContracts' "
 
-    private static void BindAmbientFieldsAsEndpoint() {
-      BindCustomEndpoint("^AmbientFields", captureAmbientFields, restoreAmbientFields);
+    private static Dictionary<string, FlowingContractDefinition> _FlowingContracts = null;
+
+    public static void DefineFlowingContract(string contractName, Action<FlowingContractDefinition> definitionMethod) {
+    
+      //safes performance during processing
+      if (contractName != contractName.ToLower()) {
+        throw new ArgumentException("The name for a FlowingContract must be in lower casing!");
+      }
+    
+      if (_FlowingContracts != null && _FlowingContracts.ContainsKey(contractName)) {
+        throw new ArgumentException($"A FlowingContract with name '{contractName}' already exisits!");
+      }
+
+      var newDefinition = new FlowingContractDefinition(false);
+      definitionMethod.Invoke(newDefinition);
+      newDefinition.MakeImmutable();
+
+      if (_FlowingContracts == null) {
+        _FlowingContracts = new Dictionary<string, FlowingContractDefinition>();
+      }
+      _FlowingContracts.Add(contractName, newDefinition);
+
     }
 
-    private static void captureAmbientFields(Action<string, string> capture) {
-      if (AmbientField.ContextAdapter == null) {
-        return;
+    private static FlowingContractDefinition GetFlowingContractDefinition(string flowingContractName) {
+      FlowingContractDefinition contract = null;
+      if (_FlowingContracts == null) {
+        if (!string.IsNullOrWhiteSpace(flowingContractName)) {
+          throw new Exception($"The '{nameof(AmbienceHub)}' was NOT YET configured to use FlowingContracts! see '{nameof(AmbienceHub.DefineFlowingContract)}'");
+        }
+        contract = FlowingContractDefinition.DefaultDefinition;
       }
-      foreach (AmbientField exposedInstance in AmbientField.ExposedInstances.Values) {
-        capture(exposedInstance.Name, exposedInstance.Value);
-      };
+      else {
+        if (string.IsNullOrWhiteSpace(flowingContractName)) {
+          throw new Exception($"The '{nameof(AmbienceHub)}' was configured to require a FlowingContract - please provide a flowingContractName when calling the capture method!");
+        }
+        if (!_FlowingContracts.TryGetValue(flowingContractName, out contract)) {
+          throw new Exception($"The '{nameof(AmbienceHub)}' cannot address the correct set of ambient data: there is no FlowingContract named '{flowingContractName}'!");
+        }
+      }
+      return contract;
     }
 
-    private static void restoreAmbientFields(IEnumerable<KeyValuePair<string, string>> sourceToRestore) {
-      if (AmbientField.ContextAdapter == null) {
-        return;
-      }
-      foreach (KeyValuePair<string, string> entryToRestore in sourceToRestore) {
-        bool restored = false;
-        foreach (AmbientField exposedInstance in AmbientField.ExposedInstances.Values) {
-          if(exposedInstance.Name.Equals(entryToRestore.Key, StringComparison.CurrentCultureIgnoreCase)) {
-            Debug.WriteLine($"{nameof(AmbienceHub)} setting value '{entryToRestore.Value}' for AmbientField '{entryToRestore.Key}'");
-            exposedInstance.Value = entryToRestore.Value; 
-            restored = true;
-            break;  
-          }
-        }
-        if (!restored) {
-          //prestage value
-          Debug.WriteLine($"{nameof(AmbienceHub)} staging value '{entryToRestore.Value}' for AmbientField '{entryToRestore.Key}'");
-          AmbientField.ContextAdapter.SetCurrentValue(entryToRestore.Key, entryToRestore.Value);
-        }
-      };
+    public static void DisableAndResetFlowingContracts() {
+      _FlowingContracts = null;
     }
 
     #endregion